fine feature added

+++ bookworm/src/main/java/com/finalassignment/bookworm/model/User.java
@@ -42,5 +42,5 @@ public class User {
     String userPassword;
 
     @Column(name = "fine")
-    Long userTotalFineAmount;
+    int userTotalFineAmount;
 }
